backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-http-client==3.3.7
sendgrid==6.11.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tzdata==2025.1
